Rename editor folder
